\name{get_data}
\alias{get_data}
\title{Get geoarchaeological and radiocarbon date data from a Google sheet}
\usage{
  get_data()
}
\value{
  A data frame
}
\description{
  This function connects to a Google sheet and brings the
  data in the the R workspace. An internet connection is
  required.
}
\examples{
my_data <- get_data()
}

