\name{psd_prep}
\alias{psd_prep}
\title{Prepares the particle size data for analysis and plotting}
\usage{
  psd_prep(my_data, ...)
}
\arguments{
  \item{my_data}{the data resulting from the
  \code{\link{get_data}} function}
}
\value{
  a data table of particle size data
}
\description{
  You don't need to run this function, it's automatically
  called by the functions for plotting and statistical
  analysis. It just does a bit of rearranging and
  simplifying.
}
\seealso{
  \code{\link{get_data}}
}

