\name{psd_three_classes}
\alias{psd_three_classes}
\title{Summarizes the particle size data into sand-silt-clay}
\usage{
  psd_three_classes(my_data, ...)
}
\arguments{
  \item{my_data}{the data resulting from the
  \code{\link{get_data}} function}
}
\value{
  a data frame
}
\description{
  You don't need to run this function, it's automatically
  called by the functions for plotting and statistical
  analysis. It just does a bit of rearranging and
  simplifying.
}
\seealso{
  \code{\link{get_data}}
}

